Fix issue 207 rebase integration
